[MOD] Cambiadas descripciones, añadidas líneas al de los xml, añadida dependencia al modulo purchase_discount, etc.
