+++ product_pricelist_rules/__openerp__.py
@@ -31,7 +31,11 @@
     "complexity": "normal",
     "summary": "",
     "description": """
-    Warning: This module is incompatible with 'product_visible_discount'
+This module extends the pricelist items, it makes possible to add a second
+discount, and to add an 'special offer' like "Buy two and get one free
+(2+1)"
+
+**Warning**: This module is incompatible with *product_visible_discount*
     """,
     "data": [
         "views/pricelist_view.xml",
