+++ product_pricelist_rules/models/__init__.py
@@ -16,4 +16,4 @@
 #
 ##############################################################################
 
-from . import pricelist
+from . import product_pricelist
