+++ product_pricelist_rules/views/pricelist_view.xml
