+++ purchase_pricelist_rules/models/purchase.py
@@ -20,8 +20,8 @@ from openerp.osv import orm, fields
 import openerp.addons.decimal_precision as dp
 
 
-class SaleOrderLineSubtotal(orm.Model):
-    _name = 'sale.order.line.subtotal'
+class PurchaseOrderLineSubtotal(orm.Model):
+    _name = 'purchase.order.line.subtotal'
 
     def _calculate_subtotal(self, cr, uid, ids, field_name, arg, context=None):
         res = {}
@@ -45,21 +45,21 @@ class SaleOrderLineSubtotal(orm.Model):
         return res
 
     _columns = {
-        'line_id': fields.many2one('sale.order.line', 'Line',
+        'line_id': fields.many2one('purchase.order.line', 'Line',
                                    ondelete='cascade'),
-        'sale_id': fields.related('line_id', 'order_id', type='many2one',
-                                  relation='sale.order', string='Sale Order',
-                                  store=True),
+        'purchase_id': fields.related('line_id', 'order_id', type='many2one',
+                                  relation='purchase.order',
+                                  string='Purchase Order', store=True),
         'item_id': fields.many2one('product.pricelist.item', 'Pricelist Item',
                                    ondelete='cascade'),
         'subtotal': fields.function(_calculate_subtotal, type='float',
                                     method=True, string='Subtotal',
-                                    obj='sale.order.pricelist.version'),
+                                    obj='purchase.order.pricelist.version'),
     }
 
 
-class SaleOrderLine(orm.Model):
-    _inherit = 'sale.order.line'
+class PurchaseOrderLine(orm.Model):
+    _inherit = 'purchase.order.line'
 
     def _calc_price_subtotal(self, cr, uid, line, context=None):
         price = line.price_unit * (1 - (line.discount or 0.0) / 100.0)
@@ -67,7 +67,7 @@ class SaleOrderLine(orm.Model):
         return price * (1 - extra_discount / 100.0)
     
     def _calc_qty(self, cr, uid, line, context=None):
-        qty = line.product_uom_qty
+        qty = line.product_qty
         if line.offer_id:
             total = line.offer_id.free_qty + line.offer_id.paid_qty
             qty = round((qty / total) * line.offer_id.paid_qty)
@@ -80,8 +80,7 @@ class SaleOrderLine(orm.Model):
         for line in self.browse(cr, uid, ids):
             new_price_subtotal = self._calc_price_subtotal(cr, uid, line)
             qty = self._calc_qty(cr, uid, line)
-
-            taxes = tax_obj.compute_all(cr, uid, line.tax_id,
+            taxes = tax_obj.compute_all(cr, uid, line.taxes_id,
                                         new_price_subtotal, qty,
                                         line.product_id,
                                         line.order_id.partner_id)
@@ -90,20 +89,26 @@ class SaleOrderLine(orm.Model):
         return res
 
     _columns = {
+        # 'discount': fields.float('Discount (%)',
+                                  # digits_compute=dp.get_precision('Discount')),
+                                  # readonly=True,
+                                  # states={'draft': [('readonly', False)]}),
         'discount2': fields.float('Discount (%)',
-                                  digits_compute=dp.get_precision('Discount'),
-                                  readonly=True,
-                                  states={'draft': [('readonly', False)]}),
+                                  digits_compute=dp.get_precision('Discount')),
+                                  # readonly=True,
+                                  # states={'draft': [('readonly', False)]}),
         'offer_id': fields.many2one('product.pricelist.item.offer', 'Offer'),
         'item_id': fields.many2one('product.pricelist.item', 'Pricelist Item'),
         'price_subtotal': fields.function(_amount_line, string='Subtotal',
                                           digits_compute=
                                           dp.get_precision('Account')),
-        'subtotal_ids': fields.one2many('sale.order.line.subtotal', 'line_id',
+        'subtotal_ids': fields.one2many('purchase.order.line.subtotal',
+                                        'line_id',
                                         'Subtotals by pricelist'),
     }
 
     _defaults = {
+        # 'discount': 0.0,
         'discount2': 0.0,
     }
 
@@ -113,13 +118,14 @@ class SaleOrderLine(orm.Model):
     ]
 
     def default_get(self, cr, uid, fields_list, context=None):
-        res = super(SaleOrderLine, self).default_get(cr, uid, fields_list,
-                                                     context=context)
+        res = super(PurchaseOrderLine, self).default_get(cr, uid, fields_list,
+                                                         context=context)
         item_obj = self.pool['product.pricelist.item']
-        item_id = item_obj.get_best_pricelist_item(cr, uid,
-                                                   context['pricelist_id'],
-                                                   context=context)
-        res.update({'item_id': item_id})
+        if context.get('pricelist_id'):
+            item_id = item_obj.get_best_pricelist_item(cr, uid,
+                                                       context['pricelist_id'],
+                                                       context=context)
+            res.update({'item_id': item_id})
         return res
 
     def onchange_item_id(self, cr, uid, ids, item_id, context=None):
@@ -134,13 +140,13 @@ class SaleOrderLine(orm.Model):
         }
         return {'value': values}
 
-class SaleOrder(orm.Model):
-    _inherit = 'sale.order'
+class PurchaseOrder(orm.Model):
+    _inherit = 'purchase.order'
 
     def _amount_line_tax(self, cr, uid, line, context=None):
         val = 0.0
         tax_obj = self.pool['account.tax']
-        line_obj = self.pool['sale.order.line']
+        line_obj = self.pool['purchase.order.line']
         new_price_subtotal = line_obj._calc_price_subtotal(cr, uid, line)
         qty = line_obj._calc_qty(cr, uid, line)
         for c in tax_obj.compute_all(cr, uid, line.tax_id, new_price_subtotal,
@@ -150,6 +156,7 @@ class SaleOrder(orm.Model):
         return val
 
     _columns = {
-        'subtotal_ids': fields.one2many('sale.order.line.subtotal', 'sale_id',
+        'subtotal_ids': fields.one2many('purchase.order.line.subtotal',
+                                        'purchase_id',
                                         'Subtotals per line by pricelist')
     }
