+++ purchase_pricelist_rules/views/purchase_pricelist_view.xml
