+++ sale_pricelist_rules/views/sale_pricelist_view.xml
