+++ sale_pricelist_rules/views/sale_view.xml
